support caret mode
